Set the TCP6 port for the VNC server
Set the TCP6 port to the same port number as the TCP port for the VNC.
This fixes an issue where the default port number is 5900 and if there
is more than one container running on a host they both try and access
this port causing a buffer overflow.

Also add some sleeps to give the X server and Xfce desktop time to start
up as mentioned in #5.

Fixes #7.